Moves getShippingMethods from cart to checkout in redux store
