+++ packages/venia-concept/src/actions/checkout/actions.js
@@ -22,6 +22,10 @@ const actionMap = {
         ACCEPT: null,
         REJECT: null
     },
+    GET_SHIPPING_METHODS: {
+        REQUEST: null,
+        RECEIVE: null
+    },
     SHIPPING_METHOD: {
         SUBMIT: null,
         ACCEPT: null,
